pyproject.toml
[project]
name = "codeatlas"
version = "0.1.0"
description = "Fork repos, auto-document them with Ollama, and publish MkDocs sites."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygit2>=1.14",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "ollama>=0.1.7",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "rich>=13.9",
]

[project.scripts]
codeatlas = "codeatlas.main:cli"
